Implement 128-bit ints where supported.
Ratio will violate the spec and select int128_t as its type.
This because there's an issue with being unable to change platform ABI,
making it impossible to change intmax_t.

Since this is a new OS, I might actually change the default anyway,
but that decision I'll leave for later.