Implement codecvt conversions.
While here, make a final decision on mbstate_t.  8 bytes is enough and
by making it a 64-bit integer, we can take advantage of the native
alignment of the CPU.

Codecvt uses __builtin_bzero and __builtin_memcpy to manipulate mbstate_t,
to prevent type punning issues while allowing the compiler to optimize.