Implement std::deque.
std::deque is implemented using a chain of small (8 kB) arrays.
An underlying circular buffer and a host of conversion iterators are
used to provide the facade of an infinitely extensible array.

Due to the underlying array hanging on to allocated storage when not in
use, it should be very fast for queue handling (crossing fingers).

Complexity:
- begin()/end() operations: O(1)
- random access (operator[]) operations: O(1)
- push_back/push_front: O(1)  (amortized O(1) for the circular_buffer)
- pop_back/pop_front: O(1)
- erase/insert from middle of array: O(n)

Most operations on the iterators should be simple arithmatic, so the
convoluted layout should have little effect on performance.  Most
operations should restrict themselves to registers only.