Implement page_cache.
The page cache is responsible for figuring out which pages to sacrifice
in case of memory shortage.

It maintains a list of cold and a list of hot pages, which are kept the same
size.  It also supports speculative loads, where a page is loaded before its
first access (those pages are considered cold, but access will promote them
to (regular) cold instead of hot).